way to only install .app

+++ create-vm-template.bash
@@ -31,22 +31,24 @@ else
   [[ -z $TEMPLATE ]] && echo "Did you specify the path to an macOS installer .app?" && exit 1
   INSTALLER_LOCATION="$1"
 fi
-cd $HOME
-# Cleanup already existing Template
-curl -s -X DELETE ${URL_PROTOCOL}$CLOUD_CONTROLLER_ADDRESS:$CLOUD_CONTROLLER_PORT/api/v1/registry/vm\?id\=$ANKA_VM_TEMPLATE_UUID &>/dev/null || true
-sudo anka delete --yes $ANKA_VM_TEMPLATE_UUID &>/dev/null || true
-sudo anka delete --yes $TEMPLATE &>/dev/null || true
-# Create Base Template
-echo "]] Creating $TEMPLATE using $INSTALLER_LOCATION ..."
-sudo anka create --ram-size 10G --cpu-count 6 --disk-size 80G --app "$INSTALLER_LOCATION" $TEMPLATE
-## Change UUID for Template
-CUR_UUID=$(sudo anka --machine-readable list | jq -r ".body[] | select(.name==\"$TEMPLATE\") | .uuid")
-sudo mv "$(sudo anka config vm_lib_dir)/$CUR_UUID" "$(sudo anka config vm_lib_dir)/$ANKA_VM_TEMPLATE_UUID"
-sudo sed -i '' "s/$CUR_UUID/$ANKA_VM_TEMPLATE_UUID/" "$(sudo anka config vm_lib_dir)/$ANKA_VM_TEMPLATE_UUID/$CUR_UUID.yaml"
-sudo mv "$(sudo anka config vm_lib_dir)/$ANKA_VM_TEMPLATE_UUID/$CUR_UUID.yaml" "$(sudo anka config vm_lib_dir)/$ANKA_VM_TEMPLATE_UUID/$ANKA_VM_TEMPLATE_UUID.yaml"
-# Add Registry to CLI (if the registry was installed locally)
-if [[ ! -z $(grep anka.registry /etc/hosts) && -z $(sudo anka registry list-repos | grep $CLOUD_REGISTRY_REPO_NAME) ]]; then
-  sudo anka registry add $CLOUD_REGISTRY_REPO_NAME ${URL_PROTOCOL}$CLOUD_REGISTRY_ADDRESS:$CLOUD_REGISTRY_PORT
-  sudo anka registry list-repos
+if [[ "$1" != "--no-anka-create" ]]; then
+  cd $HOME
+  # Cleanup already existing Template
+  curl -s -X DELETE ${URL_PROTOCOL}$CLOUD_CONTROLLER_ADDRESS:$CLOUD_CONTROLLER_PORT/api/v1/registry/vm\?id\=$ANKA_VM_TEMPLATE_UUID &>/dev/null || true
+  sudo anka delete --yes $ANKA_VM_TEMPLATE_UUID &>/dev/null || true
+  sudo anka delete --yes $TEMPLATE &>/dev/null || true
+  # Create Base Template
+  echo "]] Creating $TEMPLATE using $INSTALLER_LOCATION ..."
+  sudo anka create --ram-size 10G --cpu-count 6 --disk-size 80G --app "$INSTALLER_LOCATION" $TEMPLATE
+  ## Change UUID for Template
+  CUR_UUID=$(sudo anka --machine-readable list | jq -r ".body[] | select(.name==\"$TEMPLATE\") | .uuid")
+  sudo mv "$(sudo anka config vm_lib_dir)/$CUR_UUID" "$(sudo anka config vm_lib_dir)/$ANKA_VM_TEMPLATE_UUID"
+  sudo sed -i '' "s/$CUR_UUID/$ANKA_VM_TEMPLATE_UUID/" "$(sudo anka config vm_lib_dir)/$ANKA_VM_TEMPLATE_UUID/$CUR_UUID.yaml"
+  sudo mv "$(sudo anka config vm_lib_dir)/$ANKA_VM_TEMPLATE_UUID/$CUR_UUID.yaml" "$(sudo anka config vm_lib_dir)/$ANKA_VM_TEMPLATE_UUID/$ANKA_VM_TEMPLATE_UUID.yaml"
+  # Add Registry to CLI (if the registry was installed locally)
+  if [[ ! -z $(grep anka.registry /etc/hosts) && -z $(sudo anka registry list-repos | grep $CLOUD_REGISTRY_REPO_NAME) ]]; then
+    sudo anka registry add $CLOUD_REGISTRY_REPO_NAME ${URL_PROTOCOL}$CLOUD_REGISTRY_ADDRESS:$CLOUD_REGISTRY_PORT
+    sudo anka registry list-repos
+  fi
+  $SCRIPT_DIR/create-vm-template-tags.bash $TEMPLATE
 fi
-$SCRIPT_DIR/create-vm-template-tags.bash $TEMPLATE
